chore(setup): setup base app for deployment
- modify existing express server to allow for specified stacks on pivotal tracker board
- create .env.sample file for tracking changes to environment variables
- create build script and set up start script to run transpiled es5 code on production
- create app.json file for heroku review apps

[Delivers #167190425]
