+++ .env.sample
@@ -1,0 +1,1 @@
+PORT = an integer representing the port through which the server will be accessed e.g. 3000
