+++ package.json
@@ -3,29 +3,30 @@
   "version": "1.0.0",
   "description": "A Social platform for the creative at heart",
   "main": "index.js",
-  "scripts": {},
+  "scripts": {
+    "start": "node build/index.js",
+    "dev": "export DEBUG=dev && nodemon --exec babel-node server/index.js",
+    "clean": "rm -rf build && mkdir build",
+    "build": "npm run clean && babel -d ./build ./server"
+  },
   "author": "Andela Simulations Programme",
   "license": "MIT",
   "dependencies": {
     "body-parser": "^1.18.3",
+    "chalk": "^2.4.2",
     "cors": "^2.8.4",
+    "debug": "^4.1.1",
     "dotenv": "^6.0.0",
-    "ejs": "^2.6.1",
     "errorhandler": "^1.5.0",
     "express": "^4.16.3",
-    "express-jwt": "^5.3.1",
-    "express-session": "^1.15.6",
     "jsonwebtoken": "^8.3.0",
-    "method-override": "^2.3.10",
-    "methods": "^1.1.2",
-    "mongoose": "^5.2.2",
-    "mongoose-unique-validator": "^2.0.1",
-    "passport": "^0.4.0",
-    "passport-local": "^1.0.0",
-    "request": "^2.87.0",
-    "underscore": "^1.9.1"
+    "morgan": "^1.9.1"
   },
   "devDependencies": {
+    "@babel/cli": "^7.5.5",
+    "@babel/core": "^7.5.5",
+    "@babel/node": "^7.5.5",
+    "@babel/preset-env": "^7.5.5",
     "eslint": "^6.1.0",
     "eslint-config-airbnb-base": "^13.2.0",
     "eslint-config-prettier": "^6.0.0",
@@ -33,6 +34,7 @@
     "eslint-plugin-prettier": "^3.1.0",
     "husky": "^3.0.1",
     "lint-staged": "^9.2.0",
+    "nodemon": "^1.19.1",
     "prettier": "^1.18.2"
   },
   "husky": {
