--- routes/index.js
@@ -1,5 +1,0 @@
-const router = require("express").Router();
-
-router.use("/api", require("./api"));
-
-module.exports = router;
